[ADD] Backend สามารถเพิ่ม Tracking ตำแหน่งของ User คนๆนั้นได้

+++ WatchmeBackend/src/main/java/com/dreamteam/watchme/model/Position.java
@@ -11,23 +11,23 @@ package com.dreamteam.watchme.model;
  */
 public class Position {
 
-    private long latitude;
+    private double latitude;
 
-    private long longitude;
+    private double longitude;
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(long latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(long longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
